backend/requirements.txt
# CULTR Ventures — Backend Dependencies
# Pin major versions for reproducibility

# ── Web Framework ───────────────────────────────────────────────────
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.32.0,<1.0
pydantic>=2.10.0,<3.0
pydantic-settings>=2.6.0,<3.0

# ── Database ────────────────────────────────────────────────────────
sqlalchemy>=2.0.36,<3.0
asyncpg>=0.30.0,<1.0
alembic>=1.14.0,<2.0
pgvector>=0.3.6,<1.0

# ── Redis / Cache ───────────────────────────────────────────────────
redis>=5.2.0,<6.0

# ── Task Queue ──────────────────────────────────────────────────────
celery>=5.4.0,<6.0

# ── Vector Store ────────────────────────────────────────────────────
qdrant-client>=1.12.0,<2.0

# ── AI / LLM ───────────────────────────────────────────────────────
anthropic>=0.42.0,<1.0
openai>=1.58.0,<2.0

# ── Payments ────────────────────────────────────────────────────────
stripe>=11.4.0,<12.0

# ── Utilities ───────────────────────────────────────────────────────
httpx>=0.28.0,<1.0
python-dotenv>=1.0.0,<2.0
pyyaml>=6.0.2,<7.0
watchfiles>=1.0.0,<2.0

# ── Monitoring ──────────────────────────────────────────────────────
prometheus-client>=0.21.0,<1.0
